Fix variables type

+++ functions.cpp
@@ -37,8 +37,8 @@ vector<string> sort(vector<string> words)
 {
     string tmp = "";
     
-    for (int i = 0;i < words.size() ;i++) {
-        for(int j = i; j < words.size(); j++) {
+    for (size_t i = 0;i < words.size() ;i++) {
+        for(size_t j = i; j < words.size(); j++) {
             if(strtolower(words[i]) > strtolower(words[j])) {
                 tmp = words[i]; 
                 words[i] = words[j];
